refactor(set-password): server enforce auth
