+++ components/SetPasswordForm.tsx
@@ -1,0 +1,102 @@
+"use client";
+
+import { useState } from "react";
+import { useRouter } from "next/navigation";
+
+export default function SetPasswordForm({ email }: { email: string }) {
+  const router = useRouter();
+  const [password, setPassword] = useState("");
+  const [confirm, setConfirm] = useState("");
+  const [error, setError] = useState<string | null>(null);
+  const [loading, setLoading] = useState(false);
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setError(null);
+
+    if (password !== confirm) {
+      setError("Passwords do not match.");
+      return;
+    }
+
+    if (password.length < 6) {
+      setError("Password must be at least 6 characters long.");
+      return;
+    }
+
+    setLoading(true);
+
+    try {
+      const res = await fetch("/api/set-password", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ password }),
+      });
+
+      const data = await res.json();
+
+      if (!res.ok) {
+        throw new Error(data.error || "Failed to set password");
+      }
+
+      router.push("/profile");
+    } catch (err: any) {
+      setError(err.message || "Failed to set password");
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  return (
+    <form onSubmit={handleSubmit} className="space-y-4">
+      <div>
+        <label htmlFor="email" className="text-sm block mb-1">
+          Email
+        </label>
+        <input
+          id="email"
+          type="email"
+          value={email}
+          disabled
+          className="w-full p-2 rounded text-black bg-gray-200"
+        />
+      </div>
+      <div>
+        <label htmlFor="password" className="sr-only">
+          Password
+        </label>
+        <input
+          id="password"
+          type="password"
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+          required
+          placeholder="Password"
+          className="w-full p-2 rounded text-black"
+        />
+      </div>
+      <div>
+        <label htmlFor="confirm" className="sr-only">
+          Confirm Password
+        </label>
+        <input
+          id="confirm"
+          type="password"
+          value={confirm}
+          onChange={(e) => setConfirm(e.target.value)}
+          required
+          placeholder="Confirm Password"
+          className="w-full p-2 rounded text-black"
+        />
+      </div>
+      {error && <p className="text-red-500">{error}</p>}
+      <button
+        type="submit"
+        disabled={loading}
+        className="bg-neon text-white w-full p-2 rounded disabled:opacity-50"
+      >
+        {loading ? "Setting..." : "Set Password"}
+      </button>
+    </form>
+  );
+}
